src/part2/services/parser/requirements.txt
kafka-python>=2.0
docling>=2.28.0
transformers>=4.48.0
